feat: add ChArUco and AprilTag support refactor(GUI): add parameters group by category and some minor visual changes refactor(Evaluator framework): add plugin dependencies. Now a plugin can require the results of other plugins to be executed fix(DetectionController): fixed some bugs, like incorrect display of boards and plugin debug results when Live fix/refactor(BusyOverlay): replaced from wxWindow to wxFrame, in order to fix the bug where it was not a working transparent overlay feat(CvRegion): add 'dilationRectangle()' and 'erosionRectangle()' refactor(CvImage): now 'toBGR()' and 'toGray()' will handle the case where the image has an alpha channel fix(CvImage): fixed a bug in 'scaleImageMax()' where the source float matrix was not correctly used fix(ICamera): fixed handling of starting grab state fix(ContrastCheck): check was not correctly computing contrast when grid is tessellated. Splitted logic from isolated patterns fix(IlluminationCheck): check was not correctly computing contrast when grid is tessellated. Splitted logic from isolated patterns refactor(FocusCheck): now contrast value is given by the ContrastCheck, making it an hard requirement fix(OverexposureCheck): fixed non-breaking exception throws when board region was out of image bounds fix(DetectionUtility): fixed a bug where debug images were painted using the same non thread-safe CvPainter object by multiple threads feat(CMakeLists.txt): add necessary elements for ArUco support feat(vcpkg.json): add necessary elements for ArUco support

+++ src/calibration/DistortionModel.cpp
@@ -102,7 +102,7 @@ DistortionModel DistortionModel::tilted(
 
 /////////////////////////////////////////////////////////
 
-int DistortionModel::distortionTypeToCoeffsCount(DistortionModel::Type distModel)
+size_t DistortionModel::distortionTypeToCoeffsCount(DistortionModel::Type distModel)
 {
     switch (distModel)
     {
@@ -188,7 +188,7 @@ DistortionModel::Type DistortionModel::type() const
     return distType_;
 }
 
-int  DistortionModel::coeffsCount() const
+size_t DistortionModel::coeffsCount() const
 {
     return distortionTypeToCoeffsCount(distType_);
 }
